+++ src/msf/mfcfc/db/dao/slices/VlanIfIdDao.java
@@ -4,6 +4,6 @@ package msf.mfcfc.db.dao.slices;
 import msf.mfcfc.common.data.VlanIfId;
 import msf.mfcfc.db.dao.AbstractCommonDao;
 
-public abstract class VlanIfIdDao extends AbstractCommonDao<VlanIfId, Integer> {
+public abstract class VlanIfIdDao extends AbstractCommonDao<VlanIfId, Long> {
 
 }
